requirements.txt
Flask~=2.1.1
opencv-python~=4.5.4.60
numpy~=1.22.1
Werkzeug~=2.0.2
Pillow~=8.4.0